Less granular plugin type
We would only use one of these at a time for a given tenant so it's
overkill to be able to configure a different module for each type of
shared index. Using a single type also paves the way for a single plugin
that can handle different SI implementations with small tweaks based on
the SI.

Simple enough change but since it impacts how this is consumed by ui-rs
it's a breaking one and so this commit also bumps the major version.